+++ module/item/sheets/skill.mjs
@@ -94,7 +94,11 @@ export class BRPSkillSheet extends ItemSheet {
     const skillName = formData['system.mainName'] || this.item.system.mainName
     if (this.item.system.specialism) {
       const specialization = formData['system.specName'] || this.item.system.specName
-      formData.name = skillName + ' (' + specialization + ')'
+      if (this.item.system.chosen) {
+        formData.name = specialization + ' (' + skillName + ')'  
+      } else {
+        formData.name = skillName + ' (' + specialization + ')'
+      }
     } else {
       formData.name = skillName
     }